refactor: 직렬화/역직렬화를 위한 기본 생성자 추가
