+++ src/main/java/subway/ui/dto/LineRequest.java
@@ -1,12 +1,13 @@
-package subway.dto;
+package subway.ui.dto;
 
-public class StationRequest {
+public class LineRequest {
+
     private String name;
 
-    public StationRequest() {
+    private LineRequest() {
     }
 
-    public StationRequest(String name) {
+    public LineRequest(String name) {
         this.name = name;
     }
