+++ src/main/java/subway/ui/dto/StationRequest.java
@@ -1,0 +1,17 @@
+package subway.ui.dto;
+
+public class StationRequest {
+
+    private String name;
+
+    private StationRequest() {
+    }
+
+    public StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
