+++ src/main/java/subway/ui/dto/StationResponse.java
@@ -1,10 +1,13 @@
-package subway.dto;
+package subway.ui.dto;
 
 import subway.domain.Station;
 
 public class StationResponse {
     private Long id;
     private String name;
+
+    private StationResponse() {
+    }
 
     public StationResponse(Long id, String name) {
         this.id = id;
